Extract container utils
